requirements.txt
absl-py==2.1.0
anyio==4.4.0
APScheduler==3.6.3
attrs==24.2.0
cachetools==4.2.2
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
cmake==3.30.2
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
decorator==5.1.1
flatbuffers==24.3.25
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jax==0.4.31
jaxlib==0.4.31
kiwisolver==1.4.5
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==12.8
pytz==2024.1
scipy==1.14.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
dlib @ file:./dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models==0.3.0
